Move shader interface definition to common place
